chore(package): 패키지명 수정 - egobook_frontent → egobook 변경
