+++ app/src/main/java/com/example/egobook/ui/home/HomeFragment.kt
@@ -5,13 +5,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.example.egobook_frontent.R
-import com.example.egobook_frontent.databinding.FragmentHomeBinding
+import androidx.navigation.fragment.findNavController
+import com.example.egobook.R
+import com.example.egobook.databinding.FragmentHomeBinding
 import kotlinx.coroutines.launch
 
 class HomeFragment : Fragment() {
