+++ app/src/main/java/com/example/egobook/ui/home/HomeViewModel.kt
@@ -1,8 +1,6 @@
 package com.example.egobook.ui.home
 
-import android.R.attr.level
 import androidx.lifecycle.ViewModel
-import com.example.egobook.R
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
