+++ app/src/main/java/com/example/egobook/ui/onboarding/adapter/OnboardingVPAdapter.kt
@@ -1,12 +1,12 @@
-package com.example.egobook_frontent.ui.onboarding.adapter
+package com.example.egobook.ui.onboarding.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.egobook_frontent.ui.onboarding.view.OnboardingFirstFragment
-import com.example.egobook_frontent.ui.onboarding.view.OnboardingSecondFragment
-import com.example.egobook_frontent.ui.onboarding.view.OnboardingThirdFragment
-import com.example.egobook_frontent.ui.onboarding.view.OnboardingFourthFragment
-import com.example.egobook_frontent.ui.onboarding.view.OnboardingFifthFragment
+import com.example.egobook.ui.onboarding.view.OnboardingFirstFragment
+import com.example.egobook.ui.onboarding.view.OnboardingSecondFragment
+import com.example.egobook.ui.onboarding.view.OnboardingThirdFragment
+import com.example.egobook.ui.onboarding.view.OnboardingFourthFragment
+import com.example.egobook.ui.onboarding.view.OnboardingFifthFragment
 
 class OnboardingVPAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
     override fun createFragment(position: Int): Fragment {
