+++ app/src/main/java/com/example/egobook/ui/shop/ItemType.kt
@@ -1,4 +1,4 @@
-package com.example.egobook_frontent.ui.shop
+package com.example.egobook.ui.shop
 
 enum class ItemType {
     BACK, SKIN, DECO_1, DECO_2, BACKGROUND;
